14 finished

+++ 14-tcp_stack/code/include/tcp_timer.h
@@ -4,10 +4,11 @@
 #include "list.h"
 
 #include <stddef.h>
+#include <sys/time.h>
 
 struct tcp_timer {
 	int type;	// time-wait: 0		retrans: 1
-	int timeout;	// in micro second
+	long long timeout;	// in micro second
 	struct list_head list;
 	int enable;
 };
